Extract GlobalEvent component kind
