+++ Elsewhere/Assets/Scripts/GlobalEvent.cs
@@ -1,0 +1,16 @@
+﻿using UnityEngine;
+using UnityEngine.Events;
+
+namespace Elsewhere {
+  public abstract class GlobalEvent: MonoBehaviour {
+    [SerializeField] UnityEvent _event = new UnityEvent();
+
+    public void AddListener(UnityAction listener) {
+      _event.AddListener(listener);
+    }
+
+    public void Invoke() {
+      _event.Invoke();
+    }
+  }
+}
