+++ Elsewhere/Assets/Scripts/GlobalEvent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 287df22628904acc975ab621e0145437
+timeCreated: 1692471748
