+++ Elsewhere/Assets/Scripts/OnGlobalsAwake.cs
@@ -3,15 +3,5 @@ using UnityEngine.Events;
 using UnityEngine.Serialization;
 
 namespace Elsewhere {
-  public class OnGlobalsAwake: MonoBehaviour {
-    [SerializeField] UnityEvent _event = new UnityEvent();
-
-    public void AddListener(UnityAction listener) {
-      _event.AddListener(listener);
-    }
-
-    public void Invoke() {
-      _event.Invoke();
-    }
-  }
+  public class OnGlobalsAwake: GlobalEvent {}
 }
